requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-anthropic
langchain-community
langchain-openai
faiss-cpu
pypdf
python-dotenv
tiktoken
openai